:sparkles: Rewrite many-many relation btn bs and ds using secondary relation
