+++ dataservice/api/biospecimen/models.py
@@ -2,9 +2,30 @@ from dataservice.extensions import db
 from dataservice.api.common.model import Base, KfId
 from dataservice.api.biospecimen_genomic_file.models import (
     BiospecimenGenomicFile)
-from dataservice.api.biospecimen_diagnosis.models import (
-    BiospecimenDiagnosis)
+from dataservice.api.diagnosis.models import Diagnosis
 from sqlalchemy.ext.associationproxy import association_proxy
+from sqlalchemy import event
+
+
+class BiospecimenDiagnosis(db.Model, Base):
+    """
+    Represents association table between biospecimen table and
+    diagnosis table. Contains all biospecimen, diagnosis combiniations.
+    :param kf_id: Unique id given by the Kid's First DCC
+    :param created_at: Time of object creation
+    :param modified_at: Last time of object modification
+    """
+    __tablename__ = 'biospecimen_diagnosis'
+    __prefix__ = 'BD'
+    __table_args__ = (db.UniqueConstraint('diagnosis_id',
+                                          'biospecimen_id'),)
+    diagnosis_id = db.Column(KfId(),
+                             db.ForeignKey('diagnosis.kf_id'),
+                             nullable=False)
+
+    biospecimen_id = db.Column(KfId(),
+                               db.ForeignKey('biospecimen.kf_id'),
+                               nullable=False)
 
 
 class Biospecimen(db.Model, Base):
@@ -41,6 +62,7 @@ class Biospecimen(db.Model, Base):
            concepts from Biological Spatial Ontology
     :param dbgap_consent_code: Consent classification code from dbgap
     """
+
     __tablename__ = 'biospecimen'
     __prefix__ = 'BS'
 
@@ -109,10 +131,114 @@ class Biospecimen(db.Model, Base):
     biospecimen_genomic_files = db.relationship(BiospecimenGenomicFile,
                                                 backref='biospecimen',
                                                 cascade='all, delete-orphan')
-    diagnoses = association_proxy(
-        'biospecimen_diagnoses', 'diagnosis',
-        creator=lambda diagnosis:
-        BiospecimenDiagnosis(diagnosis=diagnosis))
-    biospecimen_diagnoses = db.relationship(BiospecimenDiagnosis,
-                                            backref='biospecimen',
-                                            cascade='all, delete-orphan')
+    diagnoses = db.relationship('Diagnosis', secondary='biospecimen_diagnosis',
+                                backref=db.backref(
+                                    'biospecimens'))
+
+
+def validate_diagnosis_biospecimen(target):
+    """
+    Ensure that both the diagnosis and biospecimen
+    have the same participant
+    If this is not the case then raise DatabaseValidationError
+    """
+    from dataservice.api.errors import DatabaseValidationError
+    # Return if biospecimen_diagnosis is None
+    if not target:
+        return
+
+    # Get biospecimen and diagnosis by id
+    bsp = None
+    ds = None
+    if target.biospecimen_id and target.diagnosis_id:
+        bsp = Biospecimen.query.get(target.biospecimen_id)
+        ds = Diagnosis.query.get(target.diagnosis_id)
+
+    # If biospecimen and diagnosis doesn't exist, return and
+    # let ORM handle non-existent foreign key
+    if bsp is None or ds is None:
+        return
+
+    # Check if this diagnosis and biospecimen refer to same participant
+    if ds.participant_id != bsp.participant_id:
+        operation = 'modify'
+        target_entity = BiospecimenDiagnosis.__tablename__
+        message = (
+            ('a diagnosis cannot be linked with a biospecimen if they '
+             'refer to different participants. diagnosis {} '
+             'refers to participant {} and '
+             'biospecimen {} refers to participant {}')
+            .format(ds.kf_id,
+                    ds.participant_id,
+                    bsp.kf_id,
+                    bsp.participant_id))
+        raise DatabaseValidationError(target_entity, operation, message)
+
+
+def validate_biospecimen(target):
+    """
+    Ensure that both the diagnosis and biospecimen
+    have the same participant
+    If this is not the case then raise DatabaseValidationError
+    """
+    from dataservice.api.errors import DatabaseValidationError
+    # Return if biospecimen is None
+    if not target:
+        return
+    print(target.diagnoses)
+    # Get diagnosis by id
+    ds = None
+    if target.diagnoses:
+        ds = Diagnosis.query.get(target.diagnoses[0].kf_id)
+
+    # If biospecimen and diagnosis doesn't exist, return and
+    # let ORM handle non-existent foreign key
+    if ds is None:
+        return
+
+    # Check if this diagnosis and biospecimen refer to same participant
+    if ds.participant_id != target.participant_id:
+        operation = 'modify'
+        target_entity = Biospecimen.__tablename__
+        message = (
+            ('a diagnosis cannot be linked with a biospecimen if they '
+             'refer to different participants. diagnosis {} '
+             'refers to participant {} and '
+             'biospecimen {} refers to participant {}')
+            .format(ds.kf_id,
+                    ds.participant_id,
+                    target.kf_id,
+                    target.participant_id))
+        raise DatabaseValidationError(target_entity, operation, message)
+
+
+@event.listens_for(BiospecimenDiagnosis, 'before_insert')
+def biospecimen_diagnosis_on_insert(mapper, connection, target):
+    """
+    Run preprocessing/validation of diagnosis before insert
+    """
+    validate_diagnosis_biospecimen(target)
+
+
+@event.listens_for(BiospecimenDiagnosis, 'before_update')
+def biospecimen_diagnosis_on_update(mapper, connection, target):
+    """
+    Run preprocessing/validation of diagnosis before update
+    """
+    validate_diagnosis_biospecimen(target)
+
+
+@event.listens_for(Biospecimen, 'before_insert')
+def biospecimen_on_insert(mapper, connection, target):
+    """
+    Run preprocessing/validation of diagnosis before insert
+    """
+    validate_biospecimen(target)
+
+
+@event.listens_for(Biospecimen, 'before_update')
+def biospecimen_on_update(mapper, connection, target):
+    """
+    Run preprocessing/validation of diagnosis before update
+    """
+    validate_biospecimen(target)
