+++ dataservice/api/biospecimen/schemas.py
@@ -42,7 +42,7 @@ class BiospecimenSchema(BaseSchema):
         exclude = (BaseSchema.Meta.exclude +
                    ('participant', 'sequencing_center') +
                    ('genomic_files', 'biospecimen_genomic_files',
-                    'biospecimen_diagnoses'))
+                    'diagnoses'))
 
     _links = ma.Hyperlinks({
         'self': ma.URLFor(Meta.resource_url, kf_id='<kf_id>'),
@@ -55,7 +55,5 @@ class BiospecimenSchema(BaseSchema):
         'diagnoses': ma.URLFor('api.diagnoses_list',
                                biospecimen_id='<kf_id>'),
         'biospecimen_genomic_files': ma.URLFor(
-            'api.biospecimen_genomic_files_list', biospecimen_id='<kf_id>'),
-        'biospecimen_diagnoses': ma.URLFor(
-            'api.biospecimen_diagnoses_list', biospecimen_id='<kf_id>')
+            'api.biospecimen_genomic_files_list', biospecimen_id='<kf_id>')
     })
