--- dataservice/api/biospecimen_diagnosis/__init__.py
@@ -1,6 +1,0 @@
-from dataservice.api.biospecimen_diagnosis.resources import (
-    BiospecimenDiagnosisAPI
-)
-from dataservice.api.biospecimen_diagnosis.resources import (
-    BiospecimenDiagnosisListAPI
-)
