+++ dataservice/api/diagnosis/models.py
@@ -1,10 +1,7 @@
-from sqlalchemy.orm import validates
 from sqlalchemy import event
 
 from dataservice.extensions import db
 from dataservice.api.common.model import Base, KfId
-from dataservice.api.biospecimen_diagnosis.models import (
-    BiospecimenDiagnosis)
 
 
 class Diagnosis(db.Model, Base):
@@ -71,6 +68,55 @@ class Diagnosis(db.Model, Base):
                                db.ForeignKey('participant.kf_id'),
                                doc='the participant who was diagnosed',
                                nullable=False)
-    biospecimen_diagnoses = db.relationship(BiospecimenDiagnosis,
-                                            backref='diagnosis',
-                                            cascade='all, delete-orphan')
+
+
+def validate_diagnosis(target):
+    """
+    Ensure that both the diagnosis and biospecimen
+    have the same participant
+    If this is not the case then raise DatabaseValidationError
+    """
+    from dataservice.api.errors import DatabaseValidationError
+    from dataservice.api.biospecimen.models import Biospecimen
+    # Return if biospecimen is None
+    if not target:
+        return
+    # Get biospecimen by id
+    bsp = None
+    if target.biospecimens:
+        bsp = Biospecimen.query.get(target.biospecimens[0].kf_id)
+    # If biospecimen and diagnosis doesn't exist, return and
+    # let ORM handle non-existent foreign key
+    if bsp is None:
+        return
+
+    # Check if this diagnosis and biospecimen refer to same participant
+    if bsp.participant_id != target.participant_id:
+        operation = 'modify'
+        target_entity = Diagnosis.__tablename__
+        message = (
+            ('a diagnosis cannot be linked with a biospecimen if they '
+             'refer to different participants. diagnosis {} '
+             'refers to participant {} and '
+             'biospecimen {} refers to participant {}')
+            .format(target.kf_id,
+                    target.participant_id,
+                    bsp.kf_id,
+                    bsp.participant_id))
+        raise DatabaseValidationError(target_entity, operation, message)
+
+
+@event.listens_for(Diagnosis, 'before_insert')
+def diagnosis_on_insert(mapper, connection, target):
+    """
+    Run preprocessing/validation of diagnosis before insert
+    """
+    validate_diagnosis(target)
+
+
+@event.listens_for(Diagnosis, 'before_update')
+def diagnosis_on_update(mapper, connection, target):
+    """
+    Run preprocessing/validation of diagnosis before update
+    """
+    validate_diagnosis(target)
